Feature: Add response slice option for GET requests
Enable the user to get a slice of the response pages, by providing a
start and a end page. This is especially useful for the get BI raw
data request as there is no other possibility to limit the amount of
pages for this request.
Enable the new functionality for the get BI raw data request.

Signed-Off-by: Martin Rudloff <rudloff.martin@googlemail.com>
Signed-Off-by: Sebastian Fricke <sebastian.fricke@posteo.net>
Reviewed-by: Sebastian Fricke <sebastian.fricke@posteo.net>